Implement catastrophe toggle logic
Change CatastropheFilter to be a Set of CatastropheType. The toggle
makes changes to this set.

The 'title' bar on the toggle ("X Évenements extrêmes") shows the
current filtered count, while the toggle shows the count on each row for
the unfiltered amount.

Note: made changes to Desktop/Mobile App components to make the build
happy, but did not test it out.