Replace thermometer with initial redesign
This differs from the figma design by showing 1990 delta temperatures,
since the avg temperatures vary too much to fix on the prepared design
(-8.2 to 14), compared to delta temperatures (-1.9 to +9.98), which we
can fit with 9 notches (-1 to 7).
With delta temperatures, it doesn't make as much sense to highlight the
+1.5C target (it's there, at 1.5), so unclear how to move forward with
the design after that.

The design is responsive, when the height changes the notches/values
adjust.

Remaining TODOs:
- emojis
- show current value
- show '1990'