Replace font with 'Matter' (#118)
Followed https://web.dev/font-best-practices/ for inlining of the font.
Used https://transfonter.org/ to convert to WOFF2/WOFF with Latin-Extended subset.

Having WOFF2+WOFF seems sufficient, based on https://transfonter.org/formats#browser-support.

Not sure what font-display strategy we want, e.g. see https://web.dev/font-best-practices/#choose-an-appropriate-font-display-strategy, but swap seems like a reasonable option:

"If displaying text in a web font is a top priority: Use font-display: swap but make sure to deliver the font early enough that it does not cause a layout shift."

Drive-by fix to move the spinner outside of content-section, otherwise it counted as a flex child and the flex gap was being applied after the second column and this somehow made it go through the main div's padding (not sure why) with the new font.