Thermometer show avg temperature (absolute) (#127)
* Thermometer show avg temperature

Showing from -3 to 4 like in the current design would not work either
because a lot of 1990 temperatures are outside that, too. I decided to
go from -6 to 10 with steps of 2 (one more notch), since that allows
representing almost all 1990 values in-bounds (-7.3 to 8.8), with some
space to have their +1.5C point not show up in the same spot (i.e. in a
later design, it won't be both the current and "+1.5C" UI elements on
the same clamped point at the extremities).
Co-authored-by: JesseEmond <“emond.jesse@gmail.com”>