Add size increase breakpoints (~responsive)
Changes the `--sz-...` CSS variables at various breakpoints, so that the
fonts and spacing increases with larger devices. A deliberate desktop
design would be better, but this makes it a bit more usable in the
meantime IMO.

Picked breakpoints based on:
https://www.freecodecamp.org/news/the-100-correct-way-to-do-css-breakpoints-88d6a5ba1862/

Some other changes (/bug-fixes) were made(/discovered) while doing this:
- border-radius is fixed, it gets its own CSS var (+ use it everywhere);
- thermometer bulb uses a sz var to be dynamic, same for zoom controls;
- emojis uses a sz var to be dynamic;
- pill min-width gets a dynamic value based on css vars, not hardcoded;
- checkbox image now stretches to full width/height;

TODOs:
- The timeline should be made a bit more responsive to be able to
grow similarly, but it is tricky as-is to get it visually nice and
probably worth its own PR.
- We should set some max-widths on some elements (e.g. input elements)
so that it looks bit nicer on desktop, maybe some left-right padding
too?