Change timeline to be responsive
Change some of the timeline styling to be based on CSS vars, so that the
breakpoint size changes apply. It doesn't look exactly the same as the
one at HEAD, but is similar.

Some of the changes:
 - Set cursor "arrow" to sz-900 instead of fixed value. This must be
   done with !important to bypass the manual style that gets set by
   vue-slider;
 - Increase sz-900 to 48px on largest breakpoint, so that the cursor is
   big enough then (rest looks okay still);
 - Change title styling to match design;
 - Use CSS vars to dynamically position some of the elements based on
   constants that change with size;
 - Remove redundant div for vue-slider tooltip (it was the same as the
   auto-added one).

Drive-by fix to leaflet zoom control styling, noticed it was off when
increasing the sz-900.