Fix bug when clicking on first padding years
The code had assumptions that padding years are never passed to
indexToYear (since they shouldn't with current vue-slider mark setup),
but this breaks apart when we call this onClick on the chart, where that
can now happen.

Changed the logic flow a bit to make it easier to work with, changed to
Math.round to snap to closest non-padding year, and added clamp for
cases where the Math.round would snap to invalid index (on the padding
before the first modeled year).