Add welcome tutorial step
This is to side-step the issue with targetting an element that depends
on data for its HTML positioning.
Currently targetting the year-selector as a place-holder for the welcome
step, but ideally would want no target. However, this changes it to be
'sticky' and changes the HTML/CSS, which breaks our styling. Have to fix
that.